FIA算子mask_copy数值溢出问题修复



# message auto-generated for no-merge-commit merge:
!2508 merge master into master

FIA算子mask_copy数值溢出问题修复

Created-by: m0_66365099
Commit-by: m0_66365099
Merged-by: turing_project1
Description: ## 描述
<!--在这里详细描述你的改动，包括改动的原因和所采取的方法。-->
修正FIA算子mask_copy跳过逻辑中的数值溢出问题
## 关联的Issue
<!-- 如果这个PR是为了解决特定的Issue，请在这里提供Issue链接。例如：关联Issue #123-->
<!-- 如果这个PR是为了解决特定的问题单，请在这里描述问题单单号。-->

## 测试
<!--描述进行了哪些测试来验证你的改动。包括但不限于二级冒烟、算子泛化等。-->

## 文档更新
<!--如果这个PR包含文档的更新，请在这里指出。例如：更新了README.md文件。-->

## 类型标签
<!--  [x] 表示选中 -->
- [ ] Bug修复
- [ ] 新特性
- [ ] 性能优化
- [ ] 文档更新
- [ ] 其他，请描述：




See merge request: cann/ops-transformer-dev!2508

+++ attention/common/op_kernel/vector_common.h
@@ -974,7 +974,7 @@ __aicore__ inline bool IsSkipAttentionmask(MaskInfo &info)
         nextToken = info.nextToken + static_cast<int64_t>(info.s2Size) - static_cast<int64_t>(info.s1Size);
     }
 
-    if (static_cast<int64_t>(info.s2StartIdx + info.s2dealNum) <= static_cast<uint64_t>(s1StartIdx) + nextToken) {
+    if (static_cast<int64_t>(info.s2StartIdx + info.s2dealNum) <= static_cast<int64_t>(s1StartIdx) + nextToken) {
         return true;
     }
     return false;
@@ -994,7 +994,7 @@ __aicore__ inline bool IsSkipAttentionmaskForPre(MaskInfo &info)
     int64_t preToken = info.preToken + static_cast<uint64_t>(info.s1Size)-static_cast<uint64_t>(info.s2Size); // 统一以左上角为原点计算Token
     int32_t s1EndIdx = info.layout == GS ? s1StartIdx + info.gs1dealNum : (info.gs1StartIdx + info.gs1dealNum) / info.gSize;
 
-    if (static_cast<int64_t>(info.s2StartIdx) + preToken >= s1EndIdx ) {
+    if (static_cast<int64_t>(info.s2StartIdx) + preToken >= static_cast<int64_t>(s1EndIdx)) {
         return true;
     }
     return false;
